Fixed issues with Player rotation
Enabled the Player object to be rotated in the editor while keeping the rotation behavior of the compass center. Overall, made elements in the player scene more robust to changes in the editor. Need to add an in-editor marker to indicate position north for the sake of correctly orienting the ship object.