Updated cloud spawn and dispel mechanic
Implemented timers for spawning and dispelling clouds while in gale mode. The dispel all command currently has no activation time. The wizard will show the action timer when trying to spawn more than the cloud max. Sometimes the timer slider bugs out and the background for the slider does not update postion in conjunction with the filler for the slider.