Implemented limits on the storm cloud
Implemented upper and lower limits for the storm cloud. If the storm cloud level does not meet the lower limit, the cloud will dispel when shifting to gale mode. When the storm cloud exceeds the upper limit, the storm cloud will dispel (explode in later iterations). The upper and lower limits are adjustable in the editor.