Implemented storm cloud charge action
The player is able to charge a storm cloud using the action button while in storm mode.  If there is no cloud one is created when the charge action executes. If there is a cloud the level of the cloud increases when the charge action executes. Need to implement a limit to how much a cloud can be charged.